backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
faster-whisper
edge-tts
pydantic
requests
